+++ lib/src/draggable.dart
@@ -57,6 +57,13 @@ class Draggable {
   /// See [Draggable] constructor.
   int minDragStartDistance;
 
+  /// The minimum distance for a drag to prevent click events.
+  /// DEPRECATED: Please use [minDragStartDistance] instead.
+  @deprecated
+  set clickSuppression(int distance) {
+    minDragStartDistance = distance;
+  }
+
   // -------------------
   // Events
   // -------------------